// libs/list-types/utiac-jr-birmingham-daily-hearing-list/src/models/types.ts
export interface UtiacJrBirminghamHearing {
  hearingTime: string;
  caseTitle: string;
  representative: string;
  caseReferenceNumber: string;
  judges: string;
  hearingType: string;
  location: string;
  additionalInformation: string;
}

export type UtiacJrBirminghamHearingList = UtiacJrBirminghamHearing[];